requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0       
autopep8==1.5.3     
boto==2.49.0        
certifi==2020.6.20  
cffi==1.14.0        
chardet==3.0.4      
discord.py==1.3.3
dnspython==1.16.0
ffmpeg-python==0.2.0
future==0.18.2
get==2019.4.13
idna==2.9
multidict==4.7.6
post==2019.4.13
public==2019.4.13
pycodestyle==2.6.0
pycparser==2.20
pymongo==3.10.1
PyNaCl==1.3.0
query-string==2019.4.13
request==2019.4.13
requests==2.24.0
six==1.15.0
toml==0.10.1
urllib3==1.25.9
websockets==8.1
yarl==1.4.2
youtube-dl==2020.6.16.1